pyproject.toml
[tool.poetry]
name = "wandb-content-navigator"
version = "0.1.0"
description = "GPT-powered naviator for Weights & Biases content"
authors = ["Morgan McGuire <morg@wandb.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
pydantic = "^2.5.3"
openai = "^1.9.0"
instructor = "^0.4.8"
uvicorn = "^0.27.0"
pandas = "^2.2.0"
fastapi = "^0.109.0"
langchain = "^0.1.6"
faiss-cpu = "^1.7.4"
python-dotenv = "^1.0.1"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"